Update AdminPanel,  Re-decentralization
Dashboard is now working, fixed the wallet balance display, made deletion simpler, added preload to AdminPanel to prevent ugliness, Delegated permissions to users and admins, wrote a brief Wallet for users,

+++ src/main/java/DAO/UserDAO.java
@@ -123,7 +123,7 @@ public class UserDAO {
         return u;
     }
 
-    public int getWallet(String username) {
+    public double getWallet(String username) {
         int temp = 0;
         DBConnect.Connect();
         if (DBConnect.isConnected()) {
